requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
langgraph==0.2.28
langchain==0.3.0
# google-genai>=0.1.0
pymupdf==1.24.10
pydantic==2.8.2
python-dotenv==1.0.1
httpx==0.27.2
google-generativeai==0.3.2

